Added a DPAPI helper
DPAPI (Data Protection API) is a Windows data encryption technology that relies on symmetric keys to encrypt data. To decrypt an encrypted piece of DPAPI data, you need to have a master key. To get a master key, you need to have a valid pre-key, which you can get from LSA secrets, or the NT hash/Password of a user.

Once you have a valid master key, you can decrypt data that was encrypted with it - you can figure out which pieces of data were encrypted with which key because each key contains a GUID that matches the GUID of the key.

Master keys can be used to decrypt Chrome cookies and login information aswell as Wi-Fi passwords.

This program, given the path to a mounted Windows drive, will extract user DPAPI master keys;